Sort And CompEdit

+++ Assets/Scripts/CompositonAdder.cs
@@ -5,36 +5,37 @@ using UnityEngine;
 public class CompositonAdder : MonoBehaviour
 {
     public GameObject buildedComp;
-
-    public Vector3 CompAreaStart;
-
+    public List<Vector3> compAreaStart;
+    
     private void Awake()
     {
         buildedComp = new GameObject();
         buildedComp.name = "Builded Comp";
-        CompAreaStart = new Vector3(0, -6, 0);
-
+        for (int i = 0; i < 9; i++)
+        {
+            compAreaStart.Add(new Vector3(i, -5.5f, 0));
+        }
     }
+    
     private void Update()
     {
         if (Input.GetMouseButtonUp(0))
         {
             GameObject selected = SendRayGameObject();
 
-            if (selected != null && CheckChampionExistOnComp(selected))
+            if (selected != null && CheckChampionExistOnComp(selected) && selected.layer == 8)
                 DeleteChampfromComp(selected);
 
 
             if (selected != null && CheckChampionExistOnComp(selected) && PlayerManager.instance.compChampions.Count != PlayerManager.instance.boardLimit)
             {
                 AddChampToComp(selected);
-
-                AddToCompArea(selected, CompAreaStart);
-                CompAreaStart.x++;
+                AddToCompArea(selected, compAreaStart);
             }
         }
     }
-    public bool CheckChampionExistOnComp(GameObject go)
+    
+    private bool CheckChampionExistOnComp(GameObject go)
     {
         if (PlayerManager.instance.compChampions.Contains(go))
             return false;
@@ -42,14 +43,14 @@ public class CompositonAdder : MonoBehaviour
         return true;
     }
 
-    public void AddChampToComp(GameObject go)
+    private void AddChampToComp(GameObject go)
     {
         PlayerManager.instance.compChampions.Add(go);
-        AddTraits(go);
+        AddTraitsToComp(go);
         UIManager.instance.RefreshBoardLimit();
     }
 
-    public void AddTraits(GameObject go)
+    private void AddTraitsToComp(GameObject go)
     {
         PlayerManager.instance.compTraits.Add(go.GetComponent<ChampionData>().traitsName1);
         PlayerManager.instance.compTraits.Add(go.GetComponent<ChampionData>().traitsName2);
@@ -58,36 +59,38 @@ public class CompositonAdder : MonoBehaviour
             PlayerManager.instance.compTraits.Add(go.GetComponent<ChampionData>().traitsName3);
     }
 
-    public void AddToCompArea(GameObject go, Vector3 Comp_pz)
+    private void AddToCompArea(GameObject go, List<Vector3> Comp_pz)
     {
-        GameObject CompChamp = Instantiate(go, Comp_pz, new Quaternion(0f, 0f, 0f, 0f)) as GameObject;
+        GameObject CompChamp = Instantiate(go, Comp_pz[0], new Quaternion(0f, 0f, 0f, 0f)) as GameObject;
         CompChamp.layer = 8;
         CompChamp.transform.parent = buildedComp.transform;
+
+        Comp_pz.RemoveAt(0);
     }
 
-    void DeleteChampfromComp(GameObject go)
+    private void DeleteChampfromComp(GameObject go)
     {
-        if (go.layer == 8)
+        compAreaStart.Insert(0, go.transform.position);
+        GameObject delete = null;
+        foreach (var item in PlayerManager.instance.compChampions)
         {
-            GameObject delete = null;
-            foreach (var item in PlayerManager.instance.compChampions)
+            if (item.GetComponent<ChampionData>().champName == go.GetComponent<ChampionData>().champName)
             {
-                if (item.GetComponent<ChampionData>().champName == go.GetComponent<ChampionData>().champName)
-                {
-                    delete = item;
-                }
+                delete = item;
             }
-            //for traits
-            if(delete != null)
-            {
-                PlayerManager.instance.compChampions.Remove(delete);
-                DestroyImmediate(go);
-                UIManager.instance.RefreshBoardLimit();
-            }
-            //PlayerManager.instance.compChampions.Remove(go);
         }
+
+        PlayerManager.instance.compTraits.Remove(go.GetComponent<ChampionData>().traitsName1);
+        PlayerManager.instance.compTraits.Remove(go.GetComponent<ChampionData>().traitsName2);
+        PlayerManager.instance.compTraits.Remove(go.GetComponent<ChampionData>().traitsName3);
+        PlayerManager.instance.compChampions.Remove(delete);
+
+        DestroyImmediate(go);
+        UIManager.instance.RefreshBoardLimit();
+        
     }
-    public GameObject SendRayGameObject()
+
+    private GameObject SendRayGameObject()
     {
         Ray ray = Camera.main.ScreenPointToRay(Input.mousePosition);
